+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 CHANGELOG
 =====
 
+# 20170808
+* General: Show ascii-art 'cheater' header on start of each output (#2)
+* Read-mode: Show cheatsheet location on displaying a cheatsheet (#3)
+* Search: If search has a match in a comment it should add the following line as well to search results (#4)
+* Updated several cheatsheets
+
+
 # 20170721
 * Added support for message_type for debugging method (0 = info, 1 = warning, 2 = error)
 * Added more debug output
 * Updated several cheatsheets
 * Search now shows a score for each cheatcheet which contains matches (#1)
+
 
 # 20170720
 * Renamed project from CHEAT to CHEATER